Reading

+++ photoniced/device/services/DeviceNoteService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
@@ -10,12 +11,38 @@ namespace photoniced.device.services
     {
         private const string InfoFileName = ".photon.json";
 
-        private static string get_file_path(string path)
+        private static string get_file_path(string path, bool create = true)
         {
             string fullPath = path + "/" + InfoFileName;
-            File.Create(fullPath).Dispose();
+            if (create == true)
+            {
+                if(!File.Exists(fullPath))
+                    File.Create(fullPath).Dispose();
+            }
 
-            return DevicePathService.get_file(path + "/" + InfoFileName).FullName;
+            try
+            {
+                return DevicePathService.get_file(fullPath).FullName;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                return null;
+            }
+
+        }
+
+        public static void delete_entry(string path, DeviceUserEntry entry)
+        {
+            string filePath = get_file_path(path, false);
+            var data = File.ReadAllText(filePath);
+            var jsonListFile = JsonConvert.DeserializeObject<List<DeviceUserEntry>>(data);
+            if (jsonListFile == null)
+                jsonListFile = new List<DeviceUserEntry>();
+            jsonListFile.Remove(entry);
+            var convertedData = JsonConvert.SerializeObject(jsonListFile, Formatting.Indented);
+
+            File.WriteAllText(filePath, convertedData);
         }
         
         public static void add_entry(string path, DeviceUserEntry entry)
@@ -30,6 +57,23 @@ namespace photoniced.device.services
 
             File.WriteAllText(filePath, convertedData);
             
+        }
+
+        public static List<DeviceUserEntry> read_entries(string path)
+        {
+            string filePath = get_file_path(path, false);
+            if (filePath == null)
+            {
+                return null;
+            }
+
+            var data = File.ReadAllText(filePath);
+            
+            var jsonListFile = JsonConvert.DeserializeObject<List<DeviceUserEntry>>(data);
+            if (jsonListFile == null)
+                jsonListFile = new List<DeviceUserEntry>();
+            
+            return jsonListFile;
         }
     }
 }
